backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "osint-tool"
version = "0.1.0"
description = "Personal OSINT tool for people search (Spain/EU focus)"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.115",
    "uvicorn[standard]>=0.30",
    "sse-starlette>=2.1",
    "pydantic>=2.7",
    "pydantic-settings>=2.4",
    "httpx[http2]>=0.27",
    "aiofiles>=24",
    "asyncpg>=0.29",
    "SQLAlchemy[asyncio]>=2.0",
    "alembic>=1.13",
    "redis>=5.0",
    "arq>=0.26",
    "jinja2>=3.1",
    "python-multipart>=0.0.9",
    "itsdangerous>=2.2",
    "phonenumbers>=8.13",
    "python-Levenshtein>=0.25",
    "rapidfuzz>=3.9",
    "anthropic>=0.40",
    "google-genai>=0.3",
    "beautifulsoup4>=4.12",
    "lxml>=5.2",
    "asyncpraw>=7.7",
    "tenacity>=9.0",
    "orjson>=3.10",
    "python-slugify>=8.0",
    "dnspython>=2.6",
    "argon2-cffi>=23.1",
    "python-dateutil>=2.9",
    "prometheus-client>=0.20",
    "structlog>=24.1",
    "openai>=1.40",
    "feedparser>=6.0",
    "imagehash>=4.3",
    "Pillow>=10.0",
    "exifread>=3.0",
    "Jinja2>=3.1",
    "scikit-learn>=1.3",
    "aiosmtplib>=3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.23",
    "respx>=0.21",
    "fakeredis>=2.23",
    "ruff>=0.5",
    "mypy>=1.10",
]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.ruff]
line-length = 110
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "S", "C4", "PTH"]
ignore = ["S101", "E501"]

[tool.mypy]
python_version = "3.11"
plugins = ["pydantic.mypy"]
